Add support for querying offline upsert tables correctly
